# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/user/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户添加",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your access Token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "forms.UserAddForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UserAddForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/user/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your access Token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "forms.UsernameForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UsernameForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/user/get": {
            "post": {
                "tags": [
                    "用户管理"
                ],
                "summary": "用户查询(单个)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your access Token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UsernameForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/user/getall": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户查询(全部)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your access Token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/user/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your access Token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "forms.UserInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "forms.UserAddForm": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "forms.UserInfo": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "forms.UsernameForm": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        }
    }
}